+++ src/components/monthview/monthview.js
@@ -336,14 +336,14 @@ MonthView.prototype = {
     if (this.isIslamic) {
       if (!s.activeDateIslamic) {
         const gregorianDate = new Date();
-        this.todayDateIslamic = this.conversions.fromGregorian(gregorianDate);
+        this.todayDateIslamic = Locale.gregorianToUmalqura(gregorianDate);
         s.activeDateIslamic = [];
         s.activeDateIslamic[0] = this.todayDateIslamic[0];
         s.activeDateIslamic[1] = this.todayDateIslamic[1];
         s.activeDateIslamic[2] = this.todayDateIslamic[2];
         year = s.activeDateIslamic[0];
         month = s.activeDateIslamic[1];
-        elementDate = this.conversions.fromGregorian(now);
+        elementDate = Locale.gregorianToUmalqura(now);
       } else {
         elementDate = s.activeDateIslamic;
       }
@@ -540,7 +540,7 @@ MonthView.prototype = {
     if (!this.currentDate) {
       if (this.isIslamic) {
         this.currentIslamicDate = [this.currentYear, this.currentMonth, this.currentDay];
-        this.currentDate = this.conversions.toGregorian(
+        this.currentDate = Locale.umalquraToGregorian(
           this.currentYear,
           this.currentMonth,
           this.currentDay
@@ -647,7 +647,7 @@ MonthView.prototype = {
    */
   firstDayOfMonth(year, month) {
     if (this.isIslamic) {
-      const firstDay = this.conversions.toGregorian(year, month, 1);
+      const firstDay = Locale.umalquraToGregorian(year, month, 1);
       return (firstDay === null ? 1 : firstDay.getDay());
     }
     return (new Date(year, month, 1)).getDay();
@@ -726,7 +726,7 @@ MonthView.prototype = {
     const min = (new Date(s.disable.minDate)).setHours(0, 0, 0, 0);
     const max = (new Date(s.disable.maxDate)).setHours(0, 0, 0, 0);
     let d2 = this.isIslamic ?
-      this.conversions.toGregorian(year, month, date) : new Date(year, month, date);
+      Locale.umalquraToGregorian(year, month, date) : new Date(year, month, date);
 
     if (!d2) {
       return false;
@@ -1184,7 +1184,7 @@ MonthView.prototype = {
   */
   selectDay(date, closePopup, insertDate) {
     if (this.isIslamic && typeof date !== 'string') {
-      this.currentIslamicDate = this.currentCalendar.conversions.fromGregorian(date);
+      this.currentIslamicDate = Locale.gregorianToUmalqura(date);
       date = stringUtils.padDate(
         this.currentIslamicDate[0],
         this.currentIslamicDate[1],
@@ -1207,7 +1207,7 @@ MonthView.prototype = {
 
     if (this.isIslamic) {
       this.currentIslamicDate = date;
-      this.currentDate = this.conversions.toGregorian(year, month, day);
+      this.currentDate = Locale.umalquraToGregorian(year, month, day);
     } else {
       this.currentDate = new Date(year, month, day);
     }
@@ -1458,7 +1458,7 @@ MonthView.prototype = {
 
         this.currentDate = firstDay;
         if (this.isIslamic) {
-          this.currentIslamicDate = this.conversions.fromGregorian(this.currentDate);
+          this.currentIslamicDate = Locale.gregorianToUmalqura(this.currentDate);
         }
         this.selectDay(this.currentDate, false, false);
       }
@@ -1483,7 +1483,7 @@ MonthView.prototype = {
 
         this.currentDate = lastDay;
         if (this.isIslamic) {
-          this.currentIslamicDate = this.conversions.fromGregorian(this.currentDate);
+          this.currentIslamicDate = Locale.gregorianToUmalqura(this.currentDate);
         }
         this.selectDay(this.currentDate, false, false);
       }
@@ -1519,7 +1519,7 @@ MonthView.prototype = {
 
         if (this.isIslamic) {
           this.currentIslamicDate = [d.year, d.month, d.day];
-          this.currentDate = this.conversions.toGregorian(
+          this.currentDate = Locale.umalquraToGregorian(
             this.currentIslamicDate[0],
             this.currentIslamicDate[1],
             this.currentIslamicDate[2]